Fix syntax lang in tests
