+++ src/main/java/com/example/teamcity/api/models/Agent.java
@@ -1,7 +1,5 @@
-package com.example.teamcity.api.models.agents;
+package com.example.teamcity.api.models;
 
-import com.example.teamcity.api.annotations.Parameterizable;
-import com.example.teamcity.api.models.BaseModel;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -13,9 +11,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class AuthorizedInfo extends BaseModel {
+public class Agent extends BaseModel {
 
-    @Parameterizable
-    private boolean status;
+    private int id;
+
+    private String name;
 
 }
