+++ src/main/java/com/example/teamcity/api/models/TestData.java
@@ -1,6 +1,5 @@
 package com.example.teamcity.api.models;
 
-import com.example.teamcity.api.models.agents.AuthorizedInfo;
 import lombok.Data;
 
 @Data
