+++ src/test/java/com/example/teamcity/ui/SetupServerTest.java
@@ -4,8 +4,10 @@ import com.example.teamcity.ui.pages.setup.FirstStartPage;
 import org.testng.annotations.Test;
 
 public class SetupServerTest extends BaseUiTest {
+
     @Test(groups = {"Setup"})
     public void setupTeamCityServerTest() {
         FirstStartPage.open().setupFirstStart();
     }
+
 }
